requirements.txt
scipy==1.0.0
bokeh==0.12.14
seaborn==0.8.1
statsmodels==0.8.0
numpy==1.12.1
pandas==0.20.3
paramiko==2.4.0
pymc3==3.3
scikit_image==0.13.0
CairoSVG==2.0.3
tqdm==4.19.4
Theano==1.0.1+2.gcd195ed
scp==0.10.2
matplotlib==2.1.2
fcsparser==0.1.3
python_frontmatter==0.4.2
skimage==0.0
PyYAML==3.12
